requirements.txt
appdirs==1.4.4
asgiref==3.3.4
asn1crypto==0.24.0
certifi==2018.8.13
cffi==1.14.5
chardet==3.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.3.2
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.23
django-admin-view-permission==1.6
django-crispy-forms==1.7.2
django-dbbackup==3.3.0
django-filter==1.1.0
djangorestframework==3.11.2
filelock==3.0.12
idna==2.7
Markdown==2.6.11
mysqlclient==2.0.3
packaging==20.9
pluggy==0.13.1
protobuf==3.6.1
py==1.10.0
pycparser==2.18
pyparsing==2.4.7
python-slugify==5.0.2
pytimeparse==1.1.8
pytz==2018.5
requests==2.20.0
simplejson==3.17.2
six==1.11.0
sqlparse==0.4.1
tabulate==0.8.9
text-unidecode==1.3
toml==0.10.2
tox==3.23.1
tqdm==4.60.0
urllib3==1.24.2
virtualenv==20.4.6
